cleaned up exports for exceptions
