+++ lib/exceptions/src/compatibility/index.ts
@@ -1,11 +1,4 @@
 
 export class VersionNotMatchedException extends Error {
-  /**
-   * Exception name.
-   *
-   * @return {string}
-   */
-  get name () : string {
-    return 'app.exceptions.compatibility.versionNotMatched'
-  }
+  //
 }
