+++ lib/exceptions/src/core/index.ts
@@ -1,11 +1,4 @@
 
 export class InvalidRouteException extends Error {
-  /**
-   * Exception name.
-   *
-   * @return {string}
-   */
-  get name () : string {
-    return 'app.exceptions.core.invalidRoute'
-  }
+  //
 }
