+++ lib/exceptions/src/global/index.ts
@@ -1,0 +1,18 @@
+
+export class Exception extends Error {
+  //
+}
+
+export class EmptyResponseException extends Error {
+  //
+}
+
+export class NotImplementedException extends Error {
+  constructor (...args: string[]) {
+    super(args.join(','))
+  }
+}
+
+export class InvalidUrlFormatException extends Error {
+  //
+}
