+++ lib/exceptions/src/management/index.ts
@@ -1,11 +1,4 @@
 
 export class PermissionNotGrantedException extends Error {
-  /**
-   * Exception name.
-   *
-   * @return {string}
-   */
-  get name () : string {
-    return 'app.exceptions.management.permissionNotGranted'
-  }
+  //
 }
