+++ lib/exceptions/src/storage/index.ts
@@ -1,12 +1,4 @@
 
 export class InvalidCookieRequestException extends Error {
-
-  /**
-   * Exception name.
-   *
-   * @return {string}
-   */
-  get name () : string {
-    return 'app.exceptions.cookies.invalidCookieRequest'
-  }
+  //
 }
