+++ lib/exceptions/src/tabs/index.ts
@@ -1,11 +1,4 @@
 
 export class TabIdUnknownException extends Error {
-  /**
-   * Exception name.
-   *
-   * @return {string}
-   */
-  get name () : string {
-    return 'app.exceptions.tabs.tabIdUnknown'
-  }
+  //
 }
